tests: refactor reused code
