+++ tests/console/conftest.py
@@ -1,63 +1,26 @@
 import os
 import re
+import shutil
 
 import pytest
 
 from cleo import ApplicationTester
+from cleo import CommandTester
 
-from poetry.console import Application as BaseApplication
-from poetry.core.masonry.utils.helpers import escape_name
-from poetry.core.masonry.utils.helpers import escape_version
-from poetry.core.packages.utils.link import Link
 from poetry.factory import Factory
-from poetry.installation.executor import Executor as BaseExecutor
+from poetry.installation import Installer
 from poetry.installation.noop_installer import NoopInstaller
 from poetry.io.null_io import NullIO
-from poetry.packages import Locker as BaseLocker
-from poetry.poetry import Poetry as BasePoetry
 from poetry.repositories import Pool
 from poetry.repositories import Repository as BaseRepository
-from poetry.repositories.exceptions import PackageNotFound
 from poetry.utils._compat import Path
 from poetry.utils.env import MockEnv
-from poetry.utils.toml_file import TomlFile
+from tests.console.helpers import Application
+from tests.console.helpers import Locker
+from tests.console.helpers import Repository
+from tests.helpers import Executor
+from tests.helpers import fixture
 from tests.helpers import mock_clone
-
-
-class Executor(BaseExecutor):
-    def __init__(self, *args, **kwargs):
-        super(Executor, self).__init__(*args, **kwargs)
-
-        self._installs = []
-        self._updates = []
-        self._uninstalls = []
-
-    @property
-    def installations(self):
-        return self._installs
-
-    @property
-    def updates(self):
-        return self._updates
-
-    @property
-    def removals(self):
-        return self._uninstalls
-
-    def _do_execute_operation(self, operation):
-        super(Executor, self)._do_execute_operation(operation)
-
-        if not operation.skipped:
-            getattr(self, "_{}s".format(operation.job_type)).append(operation.package)
-
-    def _execute_install(self, operation):
-        return 0
-
-    def _execute_update(self, operation):
-        return 0
-
-    def _execute_remove(self, operation):
-        return 0
 
 
 @pytest.fixture()
@@ -114,93 +77,6 @@ def setup(mocker, installer, installed, config, env):
     os.environ.update(environ)
 
 
-class Application(BaseApplication):
-    def __init__(self, poetry):
-        super(Application, self).__init__()
-
-        self._poetry = poetry
-
-    def reset_poetry(self):
-        poetry = self._poetry
-        self._poetry = Factory().create_poetry(self._poetry.file.path.parent)
-        self._poetry.set_pool(poetry.pool)
-        self._poetry.set_config(poetry.config)
-        self._poetry.set_locker(
-            Locker(poetry.locker.lock.path, self._poetry.local_config)
-        )
-
-
-class Locker(BaseLocker):
-    def __init__(self, lock, local_config):
-        self._lock = TomlFile(lock)
-        self._local_config = local_config
-        self._lock_data = None
-        self._content_hash = self._get_content_hash()
-        self._locked = False
-        self._lock_data = None
-        self._write = False
-
-    def write(self, write=True):
-        self._write = write
-
-    def is_locked(self):
-        return self._locked
-
-    def locked(self, is_locked=True):
-        self._locked = is_locked
-
-        return self
-
-    def mock_lock_data(self, data):
-        self.locked()
-
-        self._lock_data = data
-
-    def is_fresh(self):
-        return True
-
-    def _write_lock_data(self, data):
-        if self._write:
-            super(Locker, self)._write_lock_data(data)
-            self._locked = True
-            return
-
-        self._lock_data = data
-
-
-class Poetry(BasePoetry):
-    def __init__(self, file, local_config, package, locker, config):
-        self._file = TomlFile(file)
-        self._package = package
-        self._local_config = local_config
-        self._locker = Locker(locker.lock.path, locker._local_config)
-        self._config = config
-
-        # Configure sources
-        self._pool = Pool()
-
-
-class Repository(BaseRepository):
-    def find_packages(
-        self, name, constraint=None, extras=None, allow_prereleases=False
-    ):
-        packages = super(Repository, self).find_packages(
-            name, constraint, extras, allow_prereleases
-        )
-        if len(packages) == 0:
-            raise PackageNotFound("Package [{}] not found.".format(name))
-        return packages
-
-    def find_links_for_package(self, package):
-        return [
-            Link(
-                "https://foo.bar/files/{}-{}-py2.py3-none-any.whl".format(
-                    escape_name(package.name), escape_version(package.version.text)
-                )
-            )
-        ]
-
-
 @pytest.fixture
 def repo(http):
     http.register_uri(
@@ -212,6 +88,55 @@ def repo(http):
 @pytest.fixture
 def project_directory():
     return "simple_project"
+
+
+@pytest.fixture
+def make_poetry(tmp_dir, repo, config):
+    def make_poetry_from_fixture_name(name):
+        src = fixture(name)
+        dst = Path(tmp_dir) / name
+        shutil.copytree(src.as_posix(), dst.as_posix())
+
+        poetry = Factory().create_poetry(dst)
+        poetry.set_locker(Locker(poetry.locker.lock.path, poetry.locker._local_config))
+        poetry.set_config(config)
+
+        pool = Pool()
+        pool.add_repository(repo)
+        poetry.set_pool(pool)
+
+        return poetry
+
+    return make_poetry_from_fixture_name
+
+
+@pytest.fixture
+def make_installer_command_tester(config, executor, env):
+    def make_tester(poetry, command, app=None):
+        if app is None:
+            app = Application(poetry)
+            app.config.set_terminate_after_run(False)
+
+        tester = CommandTester(app.find(command))
+
+        executor._io = tester.io
+
+        installer = Installer(
+            tester.io,
+            env,
+            poetry.package,
+            poetry.locker,
+            poetry.pool,
+            config,
+            executor=executor,
+        )
+        installer.use_executor(True)
+        tester._command.set_installer(installer)
+        tester._command.set_env(env)
+
+        return tester
+
+    yield make_tester
 
 
 @pytest.fixture
