+++ tests/console/helpers.py
@@ -1,0 +1,84 @@
+from poetry.console import Application as BaseApplication
+from poetry.core.masonry.utils.helpers import escape_name
+from poetry.core.masonry.utils.helpers import escape_version
+from poetry.core.packages import Link
+from poetry.factory import Factory
+from poetry.packages import Locker as BaseLocker
+from poetry.repositories import Repository as BaseRepository
+from poetry.repositories.exceptions import PackageNotFound
+from poetry.utils.toml_file import TomlFile
+
+
+class Application(BaseApplication):
+    def __init__(self, poetry):
+        super(Application, self).__init__()
+
+        self._poetry = poetry
+
+    def reset_poetry(self):
+        poetry = self._poetry
+        self._poetry = Factory().create_poetry(self._poetry.file.path.parent)
+        self._poetry.set_pool(poetry.pool)
+        self._poetry.set_config(poetry.config)
+        self._poetry.set_locker(
+            Locker(poetry.locker.lock.path, self._poetry.local_config)
+        )
+
+
+class Locker(BaseLocker):
+    def __init__(self, lock, local_config):
+        self._lock = TomlFile(lock)
+        self._local_config = local_config
+        self._lock_data = None
+        self._content_hash = self._get_content_hash()
+        self._locked = False
+        self._lock_data = None
+        self._write = False
+
+    def write(self, write=True):
+        self._write = write
+
+    def is_locked(self):
+        return self._locked
+
+    def locked(self, is_locked=True):
+        self._locked = is_locked
+
+        return self
+
+    def mock_lock_data(self, data):
+        self.locked()
+
+        self._lock_data = data
+
+    def is_fresh(self):
+        return True
+
+    def _write_lock_data(self, data):
+        if self._write:
+            super(Locker, self)._write_lock_data(data)
+            self._locked = True
+            return
+
+        self._lock_data = data
+
+
+class Repository(BaseRepository):
+    def find_packages(
+        self, name, constraint=None, extras=None, allow_prereleases=False
+    ):
+        packages = super(Repository, self).find_packages(
+            name, constraint, extras, allow_prereleases
+        )
+        if len(packages) == 0:
+            raise PackageNotFound("Package [{}] not found.".format(name))
+        return packages
+
+    def find_links_for_package(self, package):
+        return [
+            Link(
+                "https://foo.bar/files/{}-{}-py2.py3-none-any.whl".format(
+                    escape_name(package.name), escape_version(package.version.text)
+                )
+            )
+        ]
